packages/aos_memory/pyproject.toml
[project]
name = "aos_memory"
version = "0.1.0"
description = "AOS Memory: Sisyphus Logic & Entropy Management"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "tiktoken",  # For token counting
    "pydantic-ai-slim[ag-ui,openai]",
    "aos_storage",
]

[tool.uv.sources]
aos_storage = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aos_memory"]
